requirements.txt
fastmcp>=0.2.0
fastapi_poe>=0.0.23
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.3
loguru>=0.7.0
uvicorn>=0.27.0
python-multipart>=0.0.7
